Use example exception instead of bad request
