+++ dev-packages/e2e-tests/test-applications/nestjs-basic/src/app.module.ts
@@ -1,22 +1,22 @@
 import { MiddlewareConsumer, Module } from '@nestjs/common';
+import { APP_FILTER } from '@nestjs/core';
 import { ScheduleModule } from '@nestjs/schedule';
 import { SentryModule } from '@sentry/nestjs/setup';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
+import { ExampleExceptionFilter } from './example.filter';
 import { ExampleMiddleware } from './example.middleware';
-import { APP_FILTER } from "@nestjs/core";
-import { BadRequestExceptionFilter } from "./bad-request.filter";
 
 @Module({
   imports: [SentryModule.forRoot(), ScheduleModule.forRoot()],
   controllers: [AppController],
   providers: [
     AppService,
-      {
-        provide: APP_FILTER,
-        useClass: BadRequestExceptionFilter
-      }
-    ],
+    {
+      provide: APP_FILTER,
+      useClass: ExampleExceptionFilter,
+    },
+  ],
 })
 export class AppModule {
   configure(consumer: MiddlewareConsumer): void {
