--- dev-packages/e2e-tests/test-applications/nestjs-basic/src/bad-request.filter.ts
@@ -1,22 +1,0 @@
-import { ArgumentsHost, BadRequestException, Catch, ExceptionFilter } from "@nestjs/common";
-import { Request, Response } from 'express';
-
-@Catch(BadRequestException)
-export class BadRequestExceptionFilter implements ExceptionFilter {
-  catch(exception: BadRequestException, host: ArgumentsHost): void {
-    const ctx = host.switchToHttp();
-    const response = ctx.getResponse<Response>();
-    const request = ctx.getRequest<Request>();
-    const status = exception.getStatus();
-    console.log('Bad request exception filter!');
-
-    response
-      .status(status)
-      .json({
-        statusCode: status,
-        timestamp: new Date().toISOString(),
-        path: request.url,
-        message: 'Bad request exception filter!',
-      });
-  }
-}
