+++ dev-packages/e2e-tests/test-applications/nestjs-basic/src/example-with-filter.exception.ts
@@ -1,0 +1,5 @@
+export class ExampleExceptionWithFilter extends Error {
+  constructor() {
+    super('Original example exception!');
+  }
+}
